+++ test/common/fixtures.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/pkg/errors"
 
-	kfenable "github.com/kubernetes-sigs/federation-v2/pkg/kubefed2/enable"
+	kfenable "github.com/kubernetes-sigs/federation-v2/pkg/kubefedctl/enable"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 )