+++ CHANGELOG.md
@@ -3,6 +3,8 @@
    kubefed2 disable now deletes the FederatedTypeConfig rather than set
    propagationEnabled, waits for the sync controller to shut down, and
    optionally removes the federated type CRD.
+ - [#825](https://github.com/kubernetes-sigs/federation-v2/pull/825) -
+   kubefed2 tool is renamed to kubefedctl.
 
 # v0.0.9
 -  [#776](https://github.com/kubernetes-sigs/federation-v2/pull/776) -